fix: clean up live tab imports
